+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `POST /poll/createTokens` Create new tokens for a poll
 * `POST /vote` Submit a ballot and consume the vote-token
 * `DELETE /poll/close` Close a poll and return the result
+* Automatically delete expired polls (a poll expire 30 days after it creation)
 
 #### Other
 * API documentation with Swagger (UI at `/swagger-ui.html`)